requirements.txt
sleekxmpp==1.1.11
python-simple-hipchat == 0.1
